SimulationConditions

The SimulationConditions module contain all required information that describes the metadata of the simulation that is unrelated to the domain (DomainDescriptors) or state (State) of the simulation.

Struct

Main.NavierStokesPropagators.SimulationConditions.SimulationCondition — Type

SimulationCondition contains all required information that describes the metadata of the simulation that is unrelated to the domain (DomainDescriptors) or state (State) of the simulation.

Public API

The method parameter are dependent on

init_cond::Char:
    The initial condition [(p) Poiseuille flow, (c) Couette flow, (r) restart flow from file].

init_kick::T:
    A floating-point value for the initial kick to the system.

force_type::Char
    Prescribes the forcing function [(m) mass flow rate, (p) pressure gradient].

force_magnitude::T
    Prescribes the forcing magnitude in relation to the forcing function.

Main.NavierStokesPropagators.SimulationConditions.planeCouetteFlow — Method

planeCouetteFlow(init_cond, init_kick)

Initialises the Plane Couette Flow configuration simulation.

Main.NavierStokesPropagators.SimulationConditions.planePoiseuilleFlow — Method

planePoiseuilleFlow(
    init_cond,
    init_kick,
    force_type,
    force_magnitude
)

Initalises the Plane Poiseuille Flow configuration for simulation.