InputOutputManagers

The InputOutputManagers module is a wrapper around the HDF5.jl I/O library, used for reading and writing results from the simulation.

Struct

Main.NavierStokesPropagators.InputOutputManagers.InputOutputManager — Type

Contains the path directory for reading and writing flowfield and grid points. Also contains the I/O frequency.

Public API

Main.NavierStokesPropagators.InputOutputManagers.write_grid — Method

write_grid(io_m, domain)

Write the wall-normal grid points to file.

Main.NavierStokesPropagators.InputOutputManagers.write_flowfield — Method

write_flowfield(io_m, state, domain)

Rearranges the flowfield state to (x,y,z) from (y,x,z) and writes to file.

Main.NavierStokesPropagators.InputOutputManagers.write_flow_statistics — Method

write_flow_statistics(io_m, state, domain)

Rearranges the statistical state to (x,y,z) from (y,x,z) and writes to file.

Main.NavierStokesPropagators.InputOutputManagers.write_attribute — Method

write_attribute(io_m, state, dt, nu)

Write the simulation attribute on the fly to a logger.

Main.NavierStokesPropagators.InputOutputManagers.read_flowfield! — Method

read_flowfield!(io_m, state)

Read the flowfield to a file and convert from (x,y,z) grid to the (y,x,z) grid.