Transformations

The Transformations module includes the definition for a Transformer, and the methods definition for the required discrete Fourier transform, and the fractional and physical grid transformation.

Struct

Main.NavierStokesPropagators.Transformations.Transformer — Type

The Transformer is an encapsulation of the data required for discrete Fourier transform via the FFTW package, containing the FFT plan and the required data for dealiasing.

Public API

The method parameter are dependent on

tf::Transformer{T}:
    the Transformer struct encapsulating the required FFTW abstractions for the discrete Fourier transform.

state::State{T}:
    the State{T} struct encapsulating the states of the simulations.

arr::Array{Complex{T},3}:
    the array to be transformed

frac_mode::Bool:
    true if the FFTW call is applied on a fractional array

dealias_mode::Bool:
    true if the 2/3-dealiasing scheme should be applied

Main.NavierStokesPropagators.Transformations.physical2fourier! — Method

physical2fourier!(tf, state)

Converts the physical representation of State arrays to frequency domain via FFT.

Main.NavierStokesPropagators.Transformations.physical2fourier! — Method

physical2fourier!(tf, arr, frac_mode, dealias_mode)

Converts the physical representation of State arrays to frequency domain via FFT for a specfic array.

Main.NavierStokesPropagators.Transformations.fourier2physical! — Method

fourier2physical!(tf, state)

Converts the frequency domain representation of State arrays to physical domain via iFFT.

Main.NavierStokesPropagators.Transformations.fourier2physical! — Method

fourier2physical!(tf, arr, frac_mode)

Converts the frequency domain representation of State arrays to physical domain via iFFT for a specfic array.

Main.NavierStokesPropagators.Transformations.y2y_F! — Method

y2y_F!(state)

Transforms the grid representation of wall-normal velocity from physical to fractional grid.

Main.NavierStokesPropagators.Transformations.y_F2y! — Method

y_F2y!(state)

Transforms the grid representation of wall-normal velocity from fractional to physical grid.